backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg[binary]==3.3.3
python-dotenv==1.0.1
pydantic==2.7.1
